i1668 scan-id from short to int

+++ ergo-wallet/src/test/scala/org/ergoplatform/wallet/utils/Generators.scala
@@ -162,7 +162,7 @@ trait Generators {
   def appStatusesGen: Gen[Set[ScanId]] = {
     if(scala.util.Random.nextBoolean()) {
       // simulate complex usage scenario
-      Gen.nonEmptyListOf(Gen.posNum[Short]).map(_.map { id: Short => ScanId @@ id }.toSet)
+      Gen.nonEmptyListOf(Gen.posNum[Int]).map(_.map { id: Int => ScanId @@ id }.toSet)
     } else {
       // simulate simple payment
       Set(PaymentsScanId)
